[feat] 메일 상세보기 페이지 구현 및 파일 업르도, 다운로드 기능 추가

+++ src/main/java/kr/co/groupworks/entity/cis/Employee.java
@@ -51,9 +51,12 @@ public class Employee {
     private LocalDateTime joinDate;
 
     @Column(name = "salary")
-    private Integer salary;
+    private Long salary;
 
     @Column(name = "supervisor_id")
-    private Integer supervisorId;
+    private Long supervisorId;
+
+//    @Column(name = "representative_image")
+//    private String representativeImage;
 
 }
